Grouped different Unit Tests in classes.
